+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed issue with returning after a participant task.
 * Added data source for membership payments.
 * Added Contact Checksum Field.
+* Added header fields to the PDF export output
 
 # Version 1.8.0